Import project, 2 days in
This is the set up for a basic brawler project but also building on the old old stuff from a game called horde shooter, it's nothing crazy but tries to be very clean and use libraries to do sophisticated things. This is sort of a modern reimagining of my melee combat game design. Heck I might make it the default project for a few things, I still want to set up blue print projects!