New spells, new mechs, art, polish, update baby
I'm not going to describe everything here tbh but it's a bunch of updates to most things in the game, new items and stuff for most categories and improved art and mechanics all around. I guess the next steps would be to give the characters actual sprites and then start messing with environment loading from room loader. That and just adding more items, more spells, more enemies, ect. It's all super easy though, that's the lovely part about good programming is you can build and build and build and if you know what you're doing and know how to scale up and keep expanding you can have 3 enemies or 50 and it's not too different. Everything's modular and clean and tucked away in inheritance and scripts... It's funny because the visuals and assets are so terrible (though improved since last update!) that the game feels extremely shoddy but under the hood it's a few days away from being a full, playable game... Maybe that's a bit of an exaggeration but I mean, tile game only took 3 weeks and I'm putting it on steam.. I'm trying to get better and faster at this whole process, really picking up speed now. Might try to make a game a month or something idk man.