Level loading simply, small adjustments to combat
So the gm room loader is set up but the transition between the rooms is ugly as sin sooo... The question becomes what can I do to fix that? I think I could set up a sort of acceptable neighbor system and make the blending between the edges of the room roughly standard so that rooms can mesh nicely just based on a fixed pattern like an autotiler. The other thing I could do is add transparency or some kind of noise to the rooms that allowed them to mesh, that sounds pretty stupid though ngl. It seems like the two robust solutions are to draw/give your rooms with an acceptable background (eg rock village can only spawn with other rock tiles around it and each tile has a connection type that forces certain neighbors, or I can set the rooms up slightly smaller than they would be other wise and have them contain the blending from the default state (say grass village has art for the border to merge but only from grass as the background/biome. ) Then these biomes could sort of contain their tiles with a slight buffer and then the background tiles could do the merging, though how to do the background tile's merge is also a question, since they're simpler I think a more rudimentary system would be fine there. It's possible the room layers could self contain borders and be depthed so that each part of the room could sort of "handle itself" over whatever background it got but I'm dubious that would work. Hmmmmm. I dunno, I'll look into it more later, some mix of biomed spawning tiles and resiliency of self to allow for small non-congruencies to be handled by the room's design with visual stylings (eg layers with borders between elements). That word salad seems to make sense. Oh yeah, the background of the map is from the Arco artist, crazy right? I love the style so much.