Jump spell, height'd combat, spells and states and pickups
A pretty hefty update here, adding logic and expanding the combat. But the big thing is I threw in the height factor for combat (and next time ai) so when you try to attack a jumping enemy it won't land unless you are both on the same plane (within the height range given by the attack) and all attacks have height and a range that you can customize per attack. I also added the ice jump effect and trails and fixed a few random structural things. Jump logic is now much more complicated, with a jump state and a falling state that generalizes all in air behavior. Other than that some minor particles and adjusting and also adding some simple pickups because I guess that makes sense for this game. Getting better I guess.