Several new maps and enemy visuals, basically assets
The maps are from the level generator (beauty engine) I've been working on and the sprites are just me putting in the work to grind out some variations on the designs with the existing frame set. Nothing crazy but cool. Also I threw in a couple small tweaks and made the "barbarians" (they're absolutely more demons at this point) always be on fire (and accompanying-ly they don't get the color filter or take damage for it)