Animation system is now spined (not spine tho)
It's not done and not applied across each character properly but the sprites are now drawn at positions and rotations based on frame arrays (i know...) instead of whole sprites. The two huge benefits of this are one, I only need to draw a few sprites for each character and everything else is automatic, so no more drawing 30 frames per character, and two, adding weapons is now trivial where before it was somewhat daunting. This system also allows for smooth stepping between animation positions if you're in to that kind of thing... But yeah, surface based composite character drawing with rigging positions. Cool.